Fix crash issue when navigating to shortlist from search
Fixed an issuew with the application crashing when user navigated from
/shortlist/edit/search back to /shortlist via the navigation bar. This
was caused by a component in the shortlist/edit/search page, Shortlist-
Container. ShortlistContainer was fetching the user's shortlist using
useQuery, but the query key that was used was only "shortlist". This same
key was being used by the useQuery hook in the Shortlist page, causing a
conflict. The fix was to change add the user's id to the query key array
in ShortlistContainer in /shortlist/edit/search.

Further optimisation should be considered with regards to fetching the
user's shortlist. Could all pages under /shortlist/edit use the same
container?