Move FilterModal handlers to global state
Moved the state values and onChange handlers for components in Filter-
Modal to zustand store. This greatly reduces prop drilling and makes the
search page much cleaner. Furthermore, previously the state of the
components was not visible if the user closed the filter modal, and for
example genre selections persisted even though visually they were not.