Add hook for retrieving watchproviders
Added  hook to be uses in the search page to retrieve a list of providers
from TMDB. This will be used to filter search results by provider. User
will also be able to search only certain providers. Given that there are
hundreds of providers, this implementation filters and keeps only the
most relevan and popular ones for the current use case.

In a more generalized application version, these should be configurable
by the user, perhaps per user or per club. For now, the current version
will suffice.