Add watchprovier filtering to search
User can now als search by a set of predetermined watchproviders. the
search page now has a set of buttons, each representing a watchprovider.
Clicking on a button will add the watchprovider to the search query and
submit a search. By default all the providers are present in the query.
This is reset the first time the user clicks one of the buttons. Further
clicks either add or remove the watchprovider from the query.