Add layout.tsx for search page
Added a layout.tsx for the search page, and moved ShortlistContainer and
the filters to it. When the user searches for movies, the whole page re-
renders. Because the ShortlistContainer and the filters were in /search
page.tsx, they were re-rendered as well. This was not only quite jarring
as everything disappeared for a second, but also causes an unnecessary
fetching of the shortlist.