Fix mobile layout margins and sizings
Tweaked the mobile layout of the app as awhole, adding bottom and side
margins to long scrollable content (such as search and watchlist), and
some buttons in the search page.

Also tweaked search page buttons sizing on smaller screens, along with
making the movie cards in scrollable content and the shortlist containers
a bit smaller.

Finally, the Filters on the search page is now sticky so that on mobile
user can still make changes to the filters without having to scroll all
the way back up.