Add selectionradio and notification to layout
When moving ShortlistContainer to edit page layout, the selection radio
and notification were forgotten and had been left commented out. This
commit adds them back in and moves them to the edit page layout so that
selection can be done in any subroute. Notification also goes away once
a selection is done.