Add sorting to /home/shortlist
Allshortlists data is now sorted in the useShortlistQuery-hook. The sorting
algorithm is simple: return 1 if the id of the shortlist is equal to the
user's id, otherwise return 0. Thus, the user's shortlist will always be
first in the list. This is paving way for more changes in the future,
where the user will interact with their list in the /home/shortlist page.