Fix hydration errors
There were somne hydration errors in watchlist and search. This fixes
these issues bymaking sure the query keys match in prefetch and client
components, along with moving loading skeleton to each routes loading.tsx.

Furthermore, there was a bug in search where if the user had not linked
their TMDB account, the getWatchlistQuery would always be pending. As
the loading skeleton was shown on the watchlist missing, this caused
users without linkage to never be able to see search results.