Fix issue with user avatar not displaying
When the user logins for the first time, their user information is saved
to the db. This includes their discord avatar image url at the time of
the first login. If the user changes their avatar, the url in the db
will not be updated, and the old avatar will be displayed. Sometimes, it
seems that the old image url will become invalid. This commit fixes the
issue by adding a check to the NextAuth signIn call, which will update
the user's avatar url if it has changed.