Fix auth error, remove movie info from card
Exporting authOptions in [...nextauth]/route.ts was causing an error, so
this was moved to a separate file.

Removed movie info from the movie card - the card is now a single image.
A button was added that takes to the movie page, where the movie info is
displayed (for now at least).