Update shortlist prefetch
Simplified how shortlists are prefetched (and fetched in general). Now,
instead of fetching the user's shortlist and all the other shortlists,
the query function fetches all shortlists and keyes them by the shortlist
id. This query is then used wherever shortlsits are needed. It is then
easy to either pick the user's shortlist from the keyed object, or render
all the shortlists except the user's. This should also simplify updating
shortlsit states with Pusher.